requirements.txt
streamlit>=1.43,<2
python-dotenv>=1.0,<2
